Refactor: Sidebar component layout으로 이동 및 SidebarLayout 추가
