+++ src/components/molecules/index.ts
@@ -2,6 +2,5 @@
 export * from './form';
 export * from './dashboardChart';
 export * from './dashboardTable';
-export * from './sidebarItem';
 export * from './table';
 export * from './input';
