+++ src/components/organisms/index.ts
@@ -1,4 +1,3 @@
 /* eslint-disable import/no-cycle */
-export * from './sidebar';
 export * from './paperContainer';
 export * from './editor';
